fix : change data type

+++ CPP/module00/ex00/megaphone.cpp
@@ -6,7 +6,7 @@ int main(int argc, char **argv)
 		std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *";
 	for (int i = 1; i < argc; ++i)
 	{
-		for (int j = 0; j < (int)strlen(argv[i]); ++j)
+		for (size_t j = 0; j < strlen(argv[i]); ++j)
 		{
 			std::cout << (char)toupper(argv[i][j]);
 		}
